+++ src/utils/Logger.ts
@@ -1,12 +1,21 @@
 import * as vs from "vscode";
 
+const configValues = ["Error", "Warning", "Info", "Debug"] as const;
+type cfgLogValues = typeof configValues[number];
+
+function toNumber(value: cfgLogValues): number {
+	return Math.abs(configValues.indexOf(value)); // if index is -1 then use the value 1 as default;
+}
+
 export class Logger {
-	static readonly INFO: "INFO" = "INFO";
-	static readonly WARN: "WARN" = "WARN";
-	static readonly ERR: "ERR" = "ERR";
+	static readonly INFO: typeof configValues[0] = configValues[0];
+	static readonly WARN: typeof configValues[1] = configValues[1];
+	static readonly ERR: typeof configValues[2] = configValues[2];
+	static readonly DEBUG: typeof configValues[3] = configValues[3];
 
 	private static loggerInstance: Logger;
 	private outputChannel: vs.OutputChannel;
+
 	private constructor() {
 		this.outputChannel = vs.window.createOutputChannel("Swagger Explorer");
 		this.outputChannel.show(true);
@@ -18,12 +27,14 @@ export class Logger {
 		return Logger.loggerInstance;
 	}
 
-	Log(
-		message: string,
-		Level: typeof Logger.INFO | typeof Logger.ERR | typeof Logger.WARN = Logger.INFO,
-		newline: boolean = false
-	) {
+	Log(message: string, Level: cfgLogValues, newline: boolean = false) {
 		const msg = `${message}`;
+		const verbosity = vs.workspace.getConfiguration("swaggerExplorer").get<cfgLogValues>("loggerVerbosity") ?? "Warning";
+		if (toNumber(verbosity) < toNumber(Level)) {
+			console.log(msg);
+			return;
+		}
+
 		try {
 			if (newline) {
 				this.outputChannel.appendLine(msg);